% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{getglobalPM25Options}
\alias{getglobalPM25Options}
\title{Return package options}
\usage{
getglobalPM25Options()
}
\value{
package options
}
\description{
Return package options
}
\examples{
\dontrun{
  getglobalPM25Options()
}

}
