% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPM_City.R
\name{getPM_City}
\alias{getPM_City}
\title{get real-timePM2.5 level at a city}
\usage{
getPM_City(cityname = "New York", radius = 50, zoomlevel = 10)
}
\arguments{
\item{cityname}{a string contains country name}

\item{radius}{miles (default = 50)}

\item{zoomlevel}{map zoom level (default = 10)}
}
\value{
a tibble
}
\description{
Query real-time  PM2.5 levels at a city  (default = San Francisco)
}
\examples{
\dontrun{getPM_City()} #require personal token
\dontrun{getPM_City(cityname = 'Beijing', radius = 50, zoomlevel = 10)}
}
