% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMatMajorWorldCities.R
\name{getPMatMajorWorldCities}
\alias{getPMatMajorWorldCities}
\title{get real-time PM2.5 level at major world cities}
\usage{
getPMatMajorWorldCities()
}
\value{
a tibble
}
\description{
Query real-time PM2.5 levels measured by US EPA's AQI (Air Quality Index) and APL (Air Pollution Level) at a list of major world cities. Return a tibble containing PM2.5 level, obervation station location, and local time. The results are sorted by PM2.5 AQI value
}
\examples{
getPMatMajorWorldCities()
}
