% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMbyCityNames.R
\name{getPMbyCityNames}
\alias{getPMbyCityNames}
\title{get real-timePM2.5 level at a city or a vector of cities}
\usage{
getPMbyCityNames(citynames = "san jose")
}
\arguments{
\item{citynames}{a vector of city names to be queryed}
}
\value{
a tibble
}
\description{
Query real-time  PM2.5 levels measured by US EPA's AQI (Air Quality Index) and APL (Air Pollution Level) at a list of cities (default "san jose"). Return a tibble containing PM2.5 level, obervation station location, and local time. The results are sorted by PM2.5 AQI
}
\examples{
getPMbyCityNames("san jose")
getPMbyCityNames(c("tokyo", "madrid"))
}
