% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMbyCountryName.R
\name{getPMbyCountryName}
\alias{getPMbyCountryName}
\title{get real-timePM2.5 level at a country}
\usage{
getPMbyCountryName(countryname = "United States", countrybound = c(50, -130,
  25, -65))
}
\arguments{
\item{countryname}{a string contains country name}

\item{countrybound}{country's geographical boundary}
}
\value{
a tibble
}
\description{
Query real-time  PM2.5 levels at a country (default = U.S.)
}
\examples{
\dontrun{getPMbyCountryName()} #require personal token
\dontrun{getPMbyCountryName(countryname = 'China', countrybound = c(55, 75, 20, 130))}
\dontrun{getPMbyCountryName(countryname = "India", countrybound = c(35, 70, 10, 90))}
\dontrun{getPMbyCountryName(countryname = "Africa", countrybound = c(20, 10, -30, 60))}
}
