% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMinRegion.R
\name{getPMinRegion}
\alias{getPMinRegion}
\title{get PM2.5 within a city region}
\usage{
getPMinRegion(cityname, d = 50)
}
\arguments{
\item{cityname}{character}

\item{d}{distance}
}
\value{
a tibble
}
\description{
get PM2.5 levels of all the stations within a city region (default distance 50 miles). return local time, station location and PM2.5 level
}
\examples{
\dontshow{setenvironment()}
getPMinRegion("newyork")
getPMinRegion("newyork", 200)
}
