\docType{package}
\name{globalPM25}
\alias{globalPM25}
\alias{globalPM25-package}
\title{Query Real-Time Local Air Pollute PM2.5 Level}
\description{
  \code{globalPM25} is a package to access real-time PM2.5. The data source is from \url{www.aqicn.org}. It provides functions \code{\link{getPMatMajorWorldCities}},
  \code{\link{getPMbyCityNames}}, \code{\link{getPMbylatlon}}, \code{\link{getPMinRegion}}, \code{\link{getPMnearme}}, \code{\link{settoken}} and \code{\link{plotPMdat}} in addition to two dataset \code{usdat} and \code{worlddat}.
}
\seealso{
  \code{\link{getPMatMajorWorldCities}}

  \code{\link{getPMbyCityNames}}

  \code{\link{getPMbylatlon}}

  \code{\link{getPMinRegion}}

  \code{\link{getPMnearme}}

  \code{\link{settoken}}

  \code{\link{plotPMdat}}
}
