% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPMdat.R
\name{plotPMdat}
\alias{plotPMdat}
\title{plot sample pm2.5 data form \code{usdat} and \code{\link{worlddat}}}
\usage{
plotPMdat()
}
\description{
plot sample pm2.5 data from stations in US and a list of selected big cities in the world
}
\examples{
plotPMdat()
}
