% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPMdata.R
\name{processPMdata}
\alias{processPMdata}
\title{process json data}
\usage{
processPMdata(mydata)
}
\arguments{
\item{mydata}{a list}
}
\value{
a tibble
}
\description{
process json data from \url{http://www.aqicn.org}
}
\keyword{internal}
