% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resetOptions}
\alias{resetOptions}
\title{Reset the global package variable \code{globalPM25.options}}
\usage{
resetOptions()
}
\value{
the default options
}
\description{
Reset the global package variable \code{globalPM25.options}
}
\examples{
\dontrun{
  resetOptions()
}

}
