% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setbaseurl}
\alias{setbaseurl}
\title{Set the root url}
\usage{
setbaseurl(baseurl)
}
\arguments{
\item{baseurl}{the value to assign}
}
\value{
the changed options
}
\description{
Set the root url
}
\examples{
\dontrun{
  setbaseurl("http://api.waqi.info")
}
}
\keyword{internal}
