% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setenvironment.R
\name{setenvironment}
\alias{setenvironment}
\title{Set an environment to access some useful constants within the package}
\usage{
setenvironment()
}
\description{
Assign some constants such as URL and API token to a new environment that can be accessed, modified and saved within the package
}
\examples{
setenvironment()
}
