% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settoken.R
\name{settoken}
\alias{settoken}
\title{Set your API toke}
\usage{
settoken(token)
}
\arguments{
\item{token}{character}
}
\description{
request your personal \code{token} from \url{http://aqicn.org/data-platform/token/}
}
\examples{
\dontshow{setenvironment()}
settoken("demo")
}
