% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setworldcities}
\alias{setworldcities}
\title{Set the city names to be queryed}
\usage{
setworldcities(citynames)
}
\arguments{
\item{citynames}{character vector of city names}
}
\value{
the changed options
}
\description{
Set the city names to be queryed
}
\examples{
\dontrun{
  setworldcities(c("tokyo", "beijing"))
}
}
