% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMbyCityNames.R
\name{getPMbyCityNames}
\alias{getPMbyCityNames}
\title{get city PM2.5 level}
\usage{
getPMbyCityNames(citynames)
}
\arguments{
\item{citynames}{a vector of city names to be queryed}
}
\value{
a tibble
}
\description{
query PM2.5 level at a list of cities. return a tibble containing Pm2.5 level, station location, local and UTC time
}
\examples{
\dontshow{setenvironment()}
getPMbyCityNames("newyork")
getPMbyCityNames(c("tokyo", "madrid"))
}
