% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMbyCityNames.R
\name{getPMbyCityNames}
\alias{getPMbyCityNames}
\title{get city PM2.5 level}
\usage{
getPMbyCityNames(citynames)
}
\arguments{
\item{citynames}{a list of city names represented in character vector to be queryed}
}
\value{
data frame containing location, time and PM2.5 level
}
\description{
get PM2.5 level at a list of cities
}
\examples{
getPMbyCityNames("newyork")
}
