% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMbyCityNames.R
\name{getPMbyCityNames}
\alias{getPMbyCityNames}
\title{get PM2.5 level at a city or a vector of cities}
\usage{
getPMbyCityNames(citynames = "san jose")
}
\arguments{
\item{citynames}{a vector of city names to be queryed}
}
\value{
a tibble
}
\description{
query PM2.5 level at a list of cities (default "san jose"). return a tibble containing Pm2.5 level, station location, local and UTC time. The results are sorted by PM2.5 levels
}
\examples{
getPMbyCityNames("san jose")
getPMbyCityNames(c("tokyo", "madrid"))
}
