% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMbyCountryName.R
\name{getPMbyCountryName}
\alias{getPMbyCountryName}
\title{get real-timePM2.5 level at a country}
\usage{
getPMbyCountryName(countryname = "usa", countrybound = c(50, -130, 25, -65))
}
\arguments{
\item{countryname}{a string contains country name}

\item{countrybound}{country's geographical boundary}
}
\value{
a tibble
}
\description{
Query real-time  PM2.5 levels at a country (default = U.S.)
}
\examples{
\dontrun{getPMbyCountryName()} #require personal token
}
