% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMinRegion.R
\name{getPMinRegion}
\alias{getPMinRegion}
\title{get real-time PM2.5 within a city region or within a geographic boundary}
\usage{
getPMinRegion(cityname, distance = 50, geobound = NULL)
}
\arguments{
\item{cityname}{character}

\item{distance}{double}

\item{geobound}{numeric vector of length 4L in the format c(lat1, lon1, lat2, lon2)}
}
\value{
a tibble
}
\description{
get real-time PM2.5 levels of all the stations within a city region (default \code{distance} 50 miles) or geographical bound in the format c(lat1, lon1, lat2, lon2). Return local time, station location and PM2.5 level. The results are sorted by PM2.5 levels
}
\examples{
\dontrun{getPMinRegion("beijing")} #require personal token
\dontrun{getPMinRegion("beijing", 200)}
\dontrun{getPMinRegion(geobound = c(38.164368, -122.85501, 37.203647, -121.431634))} #san francisco bay area
}
