% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMinRegion.R
\name{getPMinRegion}
\alias{getPMinRegion}
\title{get real-time PM2.5 within a geographic boundary}
\usage{
getPMinRegion(centroid, distance = 50, geobound)
}
\arguments{
\item{centroid}{geographical location represented by latitude and longitude}

\item{distance}{double}

\item{geobound}{numeric vector of length 4L in the format c(lat1, lon1, lat2, lon2)}
}
\value{
a tibble
}
\description{
get real-time PM2.5 levels of all the stations with a geographical boundary in the format of (a) centroid = c(lat, lon) and distance (default = 50 miles) or (b) geobound = c(lat1, lon1, lat2, lon2). Return local time, station location and PM2.5 level. The results are sorted by PM2.5 levels
}
\examples{
\dontrun{getPMinRegion(geobound = c(38.164368, -122.85501, 37.203647, -121.431634))} #san francisco bay area
}
