% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setenvironment.R
\name{setenvironment}
\alias{setenvironment}
\title{Set an environment to access useful constants that are shared within the package}
\usage{
setenvironment()
}
\description{
Assign constants such as URL and API token to a new environment that can be accessed and modified
}
\examples{
setenvironment()
}
