% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{settoken}
\alias{settoken}
\title{Set the token}
\usage{
settoken(token)
}
\arguments{
\item{token}{the value to assign}
}
\value{
the changed options
}
\description{
request your peronal token from \url{http://aqicn.org/data-platform/token/#/}
}
\examples{
\dontrun{
  settoken("demo")
}

}
