% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settoken.R
\name{settoken}
\alias{settoken}
\title{Set your API toke}
\usage{
settoken(atoken)
}
\arguments{
\item{atoken}{character vector}
}
\description{
request your token from [aqicn.org](http://aqicn.org/data-platform/token/#/)
}
\examples{
settoken("demo")
}
